Code review changes.
